refactor(state): change viewport_height from u16 to usize
Eliminates `as usize` casts at all 5 usage sites. The conversion
from u16 now happens once in ui.rs where the ratatui height is set.

+++ src/app/handler/preview.rs
@@ -25,7 +25,7 @@ pub fn handle_preview_action(
 ) {
     use crate::action::PreviewAction;
 
-    let viewport_height = state.viewport_height as usize;
+    let viewport_height = state.viewport_height;
 
     match action {
         PreviewAction::ScrollDown => {
